requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
graphdatascience==1.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.2
idna==3.10
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-experimental==0.3.2
langchain-mistralai==0.2.0
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.135
marshmallow==3.22.0
multidict==6.1.0
multimethod==1.12
mypy-extensions==1.0.0
neo4j==5.24.0
numpy==1.26.4
openai==1.51.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
propcache==0.2.0
py==1.11.0
pyarrow==16.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pypdf==5.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==8.5.0
textdistance==4.6.3
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.15.3
